requirements.txt
certifi
h5py~=3.6.0
imgaug~=0.4.0
numpy~=1.21.4
Pillow~=8.4.0
pytorch-lightning
# segmentation-models-pytorch
# Install latest version from source:
git+https://github.com/qubvel/segmentation_models.pytorch
torch==1.10.0+cu111
torchvision~=0.8.2
torchsummaryX~=1.3.0
# albumentations
git+https://github.com/albumentations-team/albumentations
# Detectron2 https://detectron2.readthedocs.io/en/latest/tutorials/install.html
-i https://dl.fbaipublicfiles.com/detectron2/wheels/cu111/torch1.10/index.html
detectron2~=0.3
# panopticapi
git+https://github.com/cocodataset/panopticapi.git
ConfigArgParse~=1.5.3
typing~=3.7.4.3
pycocotools~=2.0.3
scikit-image~=0.19.0
scipy~=1.7.3
opencv-python~=4.5.4.60
tqdm~=4.62.3
requests~=2.26.0
multiscaledeformableattention~=1.0
pytorch3d~=0.0.1
matplotlib~=3.5.0
wandb~=0.12.9
fvcore~=0.1.5.post20211023